requirements.txt
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anthropic==0.30.1
anyio==4.4.0
appnope==0.1.4
asgiref==3.8.1
assemblyai==0.26.0
astrapy==1.3.1
asttokens==2.4.1
asyncer==0.0.5
attrs==23.2.0
Authlib==1.3.1
backoff==2.2.1
bce-python-sdk==0.9.17
bcrypt==4.0.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.140
botocore==1.34.140
bson==0.5.10
build==1.2.1
cachetools==5.3.3
cassandra-driver==3.29.1
cassio==0.1.8
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
clevercsv==0.8.2
click==8.1.7
cohere==5.5.8
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
cryptography==42.0.8
dataclasses-json==0.6.7
datasets==2.19.2
debugpy==1.8.2
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docstring-parser==0.15
dspy-ai==2.4.11
duckdb==1.0.0
ecdsa==0.19.0
elastic-transport==8.13.1
elasticsearch==8.14.0
email_validator==2.2.0
emoji==2.12.1
executing==2.0.1
faiss-cpu==1.8.0.post1
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.4
filelock==3.15.4
filetype==1.2.0
firecrawl-py==0.0.16
FLAML==2.1.2
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.58.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.7.1
google-resumable-media==2.7.1
google_search_results==2.4.2
googleapis-common-protos==1.63.2
gotrue==2.5.4
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-health-checking==1.62.2
grpcio-status==1.62.2
grpcio-tools==1.62.2
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
ijson==3.3.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.3.2
jq==1.7.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kubernetes==30.1.0
langchain==0.2.6
langchain-anthropic==0.1.19
langchain-astradb==0.3.3
langchain-aws==0.1.9
langchain-chroma==0.1.2
langchain-cohere==0.1.8
langchain-community==0.2.6
langchain-core==0.2.11
langchain-experimental==0.0.61
langchain-google-genai==1.0.7
langchain-google-vertexai==1.0.6
langchain-groq==0.1.6
langchain-mistralai==0.1.9
langchain-mongodb==0.1.6
langchain-openai==0.1.14
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langdetect==1.0.9
langflow==1.0.6
langflow-base==0.0.82
langfuse==2.38.0
langsmith==0.1.84
litellm==1.41.11
loguru==0.7.2
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
metal_sdk==2.5.1
metaphor-python==0.1.23
minijinja==2.0.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.35.10
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
optuna==3.6.1
ordered-set==4.1.0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parameterized==0.9.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pgvector==0.2.5
pillow==10.4.0
pinecone-client==3.2.2
platformdirs==4.2.2
portalocker==2.10.0
postgrest==0.16.8
posthog==3.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen==0.2.32
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.2
pypdf==4.2.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.7
python-pptx==0.6.23
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qdrant-client==1.10.0
qianfan==0.3.5
rapidfuzz==3.9.4
realtime==1.0.6
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retell-sdk==4.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.2
sentry-sdk==2.8.0
setuptools==70.2.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlmodel==0.0.18
stack-data==0.6.3
starlette==0.37.2
storage3==0.7.6
StrEnum==0.4.15
structlog==24.2.0
supabase==2.5.1
supafunc==0.4.6
sympy==1.12.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
types-cachetools==5.3.0.7
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uncurl==0.0.11
unstructured==0.14.9
unstructured-client==0.23.9
upstash-vector==0.4.0
uritemplate==4.1.1
urllib3==2.2.2
uuid6==2024.1.12
uvicorn==0.30.1
uvloop==0.19.0
validators==0.28.3
watchfiles==0.22.0
wcwidth==0.2.13
weaviate-client==4.6.5
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wikipedia==1.4.0
wrapt==1.16.0
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
youtube-transcript-api==0.6.2
zep-python==2.0.0rc6
zipp==3.19.2
